Trait slos_hal::SystemKmainHooks

pub trait SystemKmainHooks {
    fn hook_kmain_loop_head(&mut self) { ... }

    fn hook_kmain_loop_inner_part(&mut self) { ... }
}

Optional kmain hook methods

Provided methods

fn hook_kmain_loop_head(&mut self)

kmain loop head hook

This is called at the beginning of every iteration of the kmain main loop.

fn hook_kmain_loop_inner_part(&mut self)

kmain inner partial loop hook

This is called after each function in the KMAIN_LOOP_PARTIALS collection.